fix(EN-1356): assign log id/date to skipped logs and replay reference conflicts
Two blocker review findings (NumaryBot on #563):

1. OrderSkipped logs landed with LedgerLog.Id=0 and Date=nil because
   ProcessOrders short-circuited before processApply's epilogue. The
   read-side index keys per-ledger logs by id, so every skip silently
   overwrote prior skips on the same ledger. assignSkipLogIDAndDate now
   allocates the slot on the PARENT scope (bypassing the overlay so the
   boundary update persists even after the overlay is dropped on rollback)
   and stamps Date from the parent. Refuses non-LedgerScoped orders and
   unknown ledgers as ErrInvalidExecutionPlan — structural invariants the
   admission whitelist already enforces, defended in depth.

2. verifySkippedOrder only checked reason ∈ whitelist; a tampered store
   could replace a successful CreatedTransaction with OrderSkipped
   (whitelisted reason) and pass. expectedSkippableOrder now also carries
   the audit-bound reference; the checker tracks references claimed by
   prior CreatedTransaction projections per ledger (referencesSeen) and
   rejects any TRANSACTION_REFERENCE_CONFLICT skip whose reference was
   never claimed earlier (CHECK_STORE_ERROR_TYPE_INVALID_SKIP). Empty
   reference and ledger-mismatch between projection envelope and
   chain-bound order are also caught.

Also fixes the lint failure that turned CI red (unconvert: unnecessary
uint32/uint64 conversions in Commit's counter replay loops).

Tests:
- processor_skippable_test.go: TestAssignSkipLogIDAndDate_* covers the
  happy path, non-LedgerScoped refusal, and unknown-ledger refusal.
- verify_skipped_order_test.go: TestVerifySkippedOrder_ReferenceConflict_*
  covers unclaimed reference, later claim, empty reference, ledger
  mismatch; TestRecordReferenceClaim_FirstSequenceWins pins the
  first-claim-wins semantic the verifier relies on.
- reference_test.go (E2E): asserts non-zero Log.Id/Date on skipped logs
  and the mixed-batch volume invariant (the skipped order's 11 units must
  not leak into the destination account's input).

+++ internal/domain/processing/overlay_scope.go
@@ -389,19 +389,19 @@ func (o *orderOverlayScope) Commit() {
 		o.Scope.PutTransactionState(k, v)
 	}
 
-	for range uint64(o.seqIDDelta) {
+	for range o.seqIDDelta {
 		o.Scope.IncrementNextSequenceID()
 	}
 
-	for range uint32(o.ledgerIDDelta) {
+	for range o.ledgerIDDelta {
 		o.Scope.IncrementNextLedgerID()
 	}
 
-	for range uint64(o.chapterIDDelta) {
+	for range o.chapterIDDelta {
 		o.Scope.IncrementNextChapterID()
 	}
 
-	for range uint64(o.queryCheckpointDelta) {
+	for range o.queryCheckpointDelta {
 		o.Scope.IncrementNextQueryCheckpointID()
 	}
 }
